Fixed issue that QDjango incorrectly handled a QString{} for a text column with NOT NULL constraint when saving
QDjango performs a data conversion for each property of a saved model. Special care must be taken for QString{} data in a property when the SQL column of the property has a NOT NULL constraint. This was accounted for with an assumption that changed from Qt 5 to Qt 6. In Qt 5 QVariant::isNull() would also return true if the variant contained an object of a builtin type with an isNull() method that returned true for that object. Now in Qt 6 QVariant::isNull() returns true if it is a null variant, false otherwise. A QString{} is null for QString but not for the containing QVariant, because it got data. Fix was to check the containing QString for isNull().